Add docs with sphinx
Add the files which are used by sphinx to automatically take the docstrings of the functions into nice formatted html files.
This will allow us to publish the documentation on read the docs.

Add the docs/_build folder to .gitignore
Add index.rst as a starting point for the documentation.
Add pyrosar.rst as a starting point for the function descriptions.

Close issue #29

Co-authored-by: Felix Cremer <felix.cremer@uni-jena.de>
Co-authored-by: Ismail Baris <ismail.baris@uni-jena.de>