Add Sentinel-1C and S1D initials
in order to be able to identify and products from the now operational
and soon coming satellites Sentinel-1C and 1D. As the data format is
assumed to be identical to the one of S1A and S1B, all that is needed
is adding its 'initials' wherever a filter for 'S1A', 'S1B' is applied.

No examples or other code is touched.